Improve RequestHandlerResult typings and doc and move it to the core package
