+++ packages/browser-crawler/package.json
@@ -59,6 +59,7 @@
         "@crawlee/types": "3.7.3",
         "@crawlee/utils": "3.7.3",
         "ow": "^0.28.1",
-        "tslib": "^2.4.0"
+        "tslib": "^2.4.0",
+        "type-fest": "^4.0.0"
     }
 }
